Refactor SConscripts of the modules
This is trying to solve a part of issue #3: "Currently, much code is
duplicated in the modules' SConscript files (searching for source
files, tests, examples, ...). This could be refactored into
site_scons to ease maintenance."

Therefore, add a site_scons/ModuleHelper.py in which all the work is
done. The modules' SConscripts call methods from this file, which
results in very short SConscripts.