Make UMFPACK solver compatible with SuiteSparse 4.4
Starting with SuiteSparse version 4.0, the type UF_long is
deprecated, SuiteSparse_long should be used instead.
With version 4.4 (which is in the packages of Ubuntu 16.04 LTS),
UF_long was removed, leading to compiler errors in UMFPACK.cpp.
Add a version switch to choose the right type in the UMFPACK solver.