Remove -DSG_* flags when compiling SG++
The idea of the SG_* defines (e.g., SG_DATADRIVEN) was that the
modules can react in their source code to which modules were
enabled on the command line and which not. For example, the pde
module could use fallback methods if it relied on methods of
datadriven, but datadriven wasn't enabled.

However, up to now, this functionality was not used by anyone.
Only pysgpp and jsgpp need the flags to determine which *.i files
and meta-headers to include. The SG_* defines had the annoying
downside that everytime when a single module was enabled or
disabled, all of the other modules would have to be recompiled
since the compiler command changed. This takes some time, and
it's simply annoying.

Therefore, remove that functionality for the regular C++ code
(but not for pysgpp/jsgpp) and wait if someone complains.