Remove legacy defines
Remove the defines -DDEFAULT_RES_THRESHOLD=-1.0 and
-DTASKS_PARALLEL_UPDOWN=4 from the default compiler command line.
Nobody seems to use them anymore. The code, in which the defines
pop up, defines the same values if the names are undefined;
thus, there should be no difference when omitting the defines in
the command line. The advantage is less clutting in the output.

If somebody would want to use different values, it can be done
via CPPDEFINES.