Let optimizers store clones of obj. functions
Previously, every optimizer class in the optimization module
got a reference to the objective function (or its gradient) and
stored that reference as a reference member variable. Apart from
the fact that storing references is error-prone and can lead to
funny results, this showed a bug in the MultiStart optimizer.

The MultiStart optimizer takes another optimizer object and
calls its optimize() function for different starting points.
It does that via OpenMP; and this is where the problem was.
In the process of getting BoundingBox to work with
OperationNaiveEval*, the operation objects now have temporary
member variables to speed up things. This means that the same
operation can't be used in parallel anymore, since this messes up
the internal state. This fact led to non-determistic test
failures in the MultiStart test in testOptimizer.

The solution implemented here is as follows: Let every optimizer
class clone every objective function in the constructor and store
unique_ptrs instead of references. This made it necessary to
implement a copy constructor for each optimizer class as having
a unique_ptr member variable implicitly deletes the implicit one.
Additionally, fix the tests which checked for, say,
getObjectiveFunction returns the exact same pointer as the
function we passed in the constructor. Check for rough function
equality instead using a Monte-Carlo approach implemented in
CheckEqualFunction.cpp.