Evaluate function in OperationNaiveEvalPart.Deriv.
Previously, in the OperationNaiveEvalPartialDerivative* classes,
only the partial derivative was evaluated, not the function
itself. This is slightly inconsistent as all the other
OperationNaiveEval* classes (Eval, EvalGradient, EvalHessian),
also evaluate the function. As the main motivation for the
PartialDerivative classes comes from my topology optimization
application, and this now also needs evaluation of the function,
I add that functionality to the PartialDerivative classes.