Fix bug in optimization/SConscript
The bug was that when checking availability of external libraries
with config.CheckLib, the library was automatically added to
the LIBS variable of the environment. This can result in wrong
linker order when compiling examples, as the SG++ modules are
appended to the existing LIBS variable (and the dependencies must
come after the SG++ modules in order to link correctly).