Remove hier. op.s for Lagrange bases
Theoretically, Lagrange bases are easier to be hierarchize.
Practically, things are not that easy as for each hierarhical
level, a linear system has still to be solved. The system should
be banded (d = 1) or at least sparse (d > 1); but for now, it's
too complicated to implement that. So, remove the Hierarchisation
operations from sgpp::base and add naive MultipleHierarchisation
operations in sgpp::optimization.