Re-enable misc tests
This reverts 0bbd45cf. That commit said there were "errors", but
not which, for what build configuration, and on what platform.
At least I'm not seeing errors on my laptop.

Disabling tests can be very dangerous. Reasons for this revert
are as follows:

1. We will probably forget that misc is currently not tested.
   GitLab issues are the proper way of handling these things.
2. 0bbd45cf disables all of the misc tests, not just the
   problematic one.
3. 0bbd45cf doesn't explain what the issue is, neither in the
   commit message nor in the code's TODO note.
4. Bring the topic back to the attention of @secklesn, who
   doesn't seem to have worked on this. (?)

If Jenkins complains about some other configuration on which
the tests fail (Windows is a highly likely candidate), then
I'm going to remove code. Preferably the faulty code in misc
if possible, and if that would have too many implications for
parallel, then the tests will be removed altogether and an issue
will be opened in GitLab.