Increase penality in AugmentedLagrangian
See #58. Maybe Lorenzo is right. I just had a problem with G04
where AugmentedLagrangian was stuck in an infeasible point,
because the penalty didn't increase. I think the line of
argumentation in Toussaint's lecture (from which I adapted
the parameters) was that in the augmented Lagrangian framework,
the Lagrange multipliers already act like penalty, and so there
wouldn't be any need to increase the penalty terms. Increasing
the penalty increase factor to 1.5 fixed my G04 problem, so let's
change the default to this value.