Disable GCC's -Wswitch-default
Clang's -Wcovered-switch-default warns when there is a default
case in a switch statement in which all possible enum cases are
already covered.

GCC's -Wswitch-default warns when there is *no* default case in a
switch statement.

I would vote in favor of Clang's warning, as having a default
case may lead to unwanted behavior if additional values are added
to the enum class. This would go silently with our current GCC
warnings.

Default cases may be useful if the number of possible values for
the switch statement is too large to check them all. However, in
this case the Clang warning would not be emitted anyway (as all
possible cases would have to be covered).

In most cases (when the function returns a value from every case),
the same behavior as a default case may be achieved by putting
the code from the default case behind the switch.