Disable replacement of stderr in cpplint.py
Disable the replacement of `sys.stderr` in `main()` since this
leads to silent deaths of the Python script. In my case, since
`SConstruct` calls the script using `python`, Python 3.x was
called. However, due to an `xrange` in the old `cpplint.py`,
a NameError was thrown, and due to the redirection of
`sys.stderr`, the error wasn't printed to the console; there was
no output at all, which looked like zero style errors...