Remove const from fuzzyinterval vectors
This is related to e29ef4b3f793108d9a1378d032827c0be3053fed
and extends the interface change from there to the Python and CPP
examples. This change is necessary to make the code work with SWIG >
4.0.

Without removing the const run into a const conversion error in the swig
generated code, due to the FuzzyInterval Director. It seems, on our end
the only possible fixes are: a) removing the const (as I have done); b)
removing the Fuzzy director (removing the possibility to inherent from
the class in python) or c) reimplement the std vector wrappers and try to
get rid of the conversion this way.

Out of those, removing the const seemed like the best choice.